requirements.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bce-python-sdk==0.9.59
cachetools==6.2.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
cssselect==1.3.0
cssutils==2.11.1
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.2
fsspec==2025.12.0
ftfy==6.3.1
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
lxml==6.0.2
MarkupSafe==3.0.3
modelscope==1.33.0
more-itertools==10.8.0
numpy==2.2.6
openai==2.14.0
opencv-contrib-python==4.10.0.84
openpyxl==3.1.5
packaging==25.0
paddleocr==3.3.2
paddlex==3.3.12
pandas==2.3.3
pillow==12.1.0
premailer==3.10.0
prettytable==3.17.0
psutil==7.2.1
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
pypdfium2==5.3.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
wcwidth==0.2.14
